[android] Add gradle build task
Introduces a new `@(GradleProjectReference)` item to help simplify the
content that needs to be added to the binding .csproj file.  This item
will translate relevant metadata to the `@(AndroidLibrary)` item that is
automatically added to the project after the gradle project is built.

A gradle build task has been added to improve msbuild output and error
logging when a gradle invocation fails.

We will now attempt to set `$ANDROID_HOME` and `$JAVA_HOME`
automatically when invoking gradle.  These values will be determined by
Android SDK tooling resolution tasks, and can be changed by setting
`$(AndroidSdkDirectory)` and `$(JavaSdkDirectory)` in the project file or
on the command line.