[macios] Add `@(XcodeProjectReference)` (#20)
Introduces a new `@(XcodeProjectReference)` item to help simplify the
content that needs to be added to the binding .csproj file.  This item
will translate relevant metadata to the `@(NativeReference)` item that
is automatically added to the project after the xcode project is built.

Build task wrappers for `xcodebuild` and `sharpie` have been added to
improve msbuild output and error logging when a tool invocation fails.