Use `bin` as default output dir for native projects (#27)
`$(GradleProjectBuildDirectory)` and `$(XcodeProjectBuildDirectory)`
props have been added to control the name of the output directory of
native builds. The default value is now `bin`, to better match .NET
project outputs.

The android and macios targets now hook into the Clean target to ensure
native project outputs are removed when Clean is ran.